requirements.txt
dramatiq==1.11.0
psycopg2-binary==2.9.1
fastapi==0.68.1
requests==2.26.0
APScheduler==3.8.1
